Added VLAN as a first class object.
Sadly the way I get the interface info, via FRR, because it is the only
way to get the interface speed easily, prevents me from getting VLAN
info associated with a port that way. So, we added another service called
vlan which runs 'bridge vlan show' and uses that output to construct the
VLAN information for the network. The JSON output of that command is too
horrible to describe and so I preferred the more compact text and used
text2fsm on it.