Add support to dump data after gather/clean
This is used for helping users provide data for us to debug a
crashing poller and also to get the data for testing the poller.
The output is either post-gather, but before extracting records
or cleanup, OR it is post-cleanup. You invoke it using the option
--run-once with either "gather" or "process" as the value. The
poller quits once data has been gathered/processed for all nodes
provided in the host list.