Fix #191: Handling connection correctly on timeout
When a command timed out, the code would automatically NULL the
connection object without closing the connection. This cuased a
new connection to be created the next time. So every timed out
command (in this case an NXOS command was always timing out on
some NXOS nodes) caused a new connection to be opened without
closing an old connection.

The fix is to not nullify a connection before closing it. And
in the case of a timed out command, there is no reason to close
the connection at all. Just move on.

asyncssh takes care of opening/closing the session correctly, as
verified by looking at the logs.