Fix #189: Handle dynamic BGP peer info correctly
NXOS provides an additional entry that contains the template for dynamic
BGP peer template as part of "show bgp all neighbors". This entry was
causing the problems reported by the bug. Each dynamic peer's info is
separately reported and that part is fine. It is just the dynamic peer
info entry that needed fixing. This patch addresses the issue by:
  * Explicitly flagging such a peer as having a state "dynamic"
  - Making the BGP entry whole by storing the first peer's convergence
    time as the establishedTime
  - Storing the dynamic peer subnet as the peer
  - Storing the num of accepted connections as v4PfxRx
  - Storing the num of acceptable connections as v4PfxTx

We also ignore this dynamic peer template from BGP assert.

Later on, we'll add support for checking if a dynamic BGP peering template
has no peers and why might that be.