Add a common build pandas query routine
Originally, the parquet engine.py had an internal model to create
a query string from whatever was passed from the UI. But with BGP,
the requirement to support displaying peerHostname meant we have to,
in the case of BGP, ignore predicate pushdown and read all the data
for a given namespace, fill out the peerHostname and then apply the
user-supplied filters. This meant BGP had to build these query filters
too.

So remove the redundancy by creating a commoon routine that both
BGP and engine.py can invoke.