Fix BGP analysis to take into account non-FRR NOS.
FRR has a bunch of advanced functionality in BGP such as unnumbered
interface support and hostname exchange capability. This allows us
to do display useful info such as peer's hostname, rely on LLDP to
find matching peers etc. All these are unfortunately missing from
most other routing suites.

This checkin addresses more of the issues arising from this difference.
We compute the peerHostname for all BGP sessions, FRR or otherwise,
if the session is Established. Furthermore, we had to handle assert
differently too. The error message is a little better now, but more
work needs to be done.

This commit also updates the tests to handle these new changes.