Remove use of eval in JSON output parser
The use of eval in python is considered a security risk. Tools like
bandit flag this use as being problematic. So fix it by removing the
use of eval and replacing with AST-based evaluation, considered the
right thing to do. In other cases, change the logic a bit to take adv
of the AST-based eval.