svcparser.py: Handle key extraction with incomplete data
In some cases, especially Junos, a key field is present, but not the
followup field, such as in the ospfNbr prefix string for Junos: vrf
field is present, and nothing after that. This caused the parser to
crash as it looked for ospf-neighbor key following the vrf. This
patch handles such a case by doing a get of the dict, rather than direct
access of the key.