Split JunOS support into QFX/MX
JunOS is not a single OS despite its name. Its commmands and outputs are different
across the different platforms such as QFX and MX. We use the model name to
differentiate between junos-mx and junos-qfx.

In connection with this change, the patch unifies the code to handle the differnt
cleaning functions for each of the OS. There was a lot of duplication of code and
an unnecessary if-elif-else logic repeated in every node's cleaning routine. We
also cleaned up some inconsistency in the names during this process.

The creaion of QFX/MX meant we had to redo some of the tests and the data gathered.