Fix JunOS interface update in poller
JunOS uses an obsoleted timestamp format to indicate the last time an
interface changed state. Since no routines exist to handle this format,
the poller uses the "x secs ago" part of the output to compute the last
change time of the interface. But this in itself is fraught with problem
since there can be jitter in using this value to compute an absolute
timestamp causing the poller to think that there's a change in state when
there's none.

To fix this, we use the obsoleted format to look for a change in status
between the previous poll record and the current one, but store the one
computed with the 'x secs ago' string.

Updated tests and data due to fixes in interface parsing bug