Update parquet writer to not save index
It appears that partitioned parquet writes do not preserve the index
of the pandas DF. In such a case, trying to save the index number of
the DF is pointless, and its also known to cause problems for some of
the other analysis engines. So, remove trying to save the index and
recreate it on read.