Partition cols are always ns/hostname
It is far more efficient to have larger parquet files than smaller ones
and after some amount of experimenting, the model of just using namespace
and hostname is sufficient to achieve the required IO efficiencies desired.
So, we fix the partition columns to be just the hostname and namespace for
all tables except sqpoller where we add the service name as an additional
partition column.