Make parquet writes efficient and performant
With pyarrow 1.0.0, a bunch of improvements have come to parquet such
as an updated on-disk format. Further, our performance study indicated
that setting the row_group_size to 100000 and using ZSTD compression
achieved the best results in terms of read/write performance as well as
on disk size. This patch has all these fixes in place.