Remove the need to modify the interface name
Previously, the interface name was one of the partition keys. In this
situation, when the interface name in some of the NOS included the "/"
character, it messed up the partitions since "/" is used as a directory
separator. To avoid this, we changed the interface name by replacing the
"/" with "-". However, we also saved the original interface name so that
we could replace it so as to make the interface name unchanged from its
original name to the user.

We changed the partition keys recently to make it such that the interface
name is no longer a partition key and so there's no need to change the
name of the interface, and there's no need to save the original interface
name and restore it.

This patch removes all that dead code and updates the tests.