Align path codebase with the other resources
path.py existed under sqobjects, which was unlike all the other
physical and logical resources (such as topology) in the code base.
This patch moves the existing path code to be under engines/pandas
like the others and creates a path object under sqobjects for external
access only, like the other resources.