Updating parquet data to handle updated FRR bgp cmd
The command to get BGP neighbors in FRR on Linux hosts was changed
but the run-once data had not changed to reflect this. As a result,
the parquet data was missing critical info causing tests to fail.
This patch addresses this by fixing the run-once command data and
thereby has updated parquet data.