Handle VRR interfaces correctly
When you look for an ARP entry for an IP, its possible to end up
with the VRR interface entry rather than the SVI interface entry.
If this happens, resolution to a remote MAC wiil fail causing the
path command to resolve incorrectly. Fix this by always resolving
to the non-VRR OIF entry in the ARPND table.