Handle new column and groupby stuff
When we do groupby, using the "observed=True" parameter is useful in
ensuring we don't display irrelevant and out-of-scope data. This required
a modification in some of the summarization routines. Also some updates
to handle the new column handling code. get_valid_df() is the hub through
which we must now call the parquet reader to ensure we get all the relevant
fields back.