Fix method to get JunOS uptime
We used the time-length/data field of the "show system uptime" command
with JSON output for JunOS. This method had a problem because of how JunOS
stops providing the seconds data. As a consequence, sq-poller continues to
update the record every polling period even though no data has changed.

We fix this by reading instead from the attributes/junos:seconds field which
provides the same info, but only in seconds.