Fix table show with specific namespaces
With the change in the parquet dir structure to include a version
of the schema as the top-level directory under a resource dir,
the function to return the list of tables for a given namespace was
broken because it relied on the older structure.

This fix addresses the issue by using the new dir structure.