Fix #250: Use ansible-inventory output only
Since the username/password or keyfiles associated with a device can
be stored in different places inside an Ansible playbook repo, and not
just in the Ansible inventory file, we switched to using the output of
ansible-inventory as input rather than the plain text of Ansible inventory
that was used before this release. Consequently we changed the option to
accept this as -a, instead of -i.

We added the parser to parse the ansible-inventory output to extract the
relevant data. We also print out more useful warnings when we realize that
the user has used a possibly wrong option to pass the inventory info i.e.
-D for ansible inventory or -a for Suzieq inventory instead of the opposite.
This has been a source of some confusion to users and so this error message
might be more helpful.