Fix #246: Junos reboot behavioral problems fixed
When JunOS rebooted, we ran into a number of issues. The format of
the downtime for a BGP neighbor changed format which caused a crash.
The interfaces/devices/ospfNbr outputs all have a similar problem
caused by the non-standard JunOS timestring. With devices for example,
the seconds granularity disappears at some point in the timestring we were
relying on. This causes the poller to continuously write out the device
data as if it was changing. To fix all these problems, we converted to using
a different field that stayed consistent through time.

This patch addresses this specific issue.