Fix #266: Make routes column logic consistent
This patch makes the columns extracted and displayed logic consistent
across both the CLI and REST API. It moves the column logic from the
show and lpm calls in CLI into the engines/pandas piece which is where
all internal column logic should now reside.

This does break existing tests which incorrectly had metric displayed,
when metric was not a desired column. So tests need to be updated too.