Fix #263: Handle default VRF correctly in address
Default VRF is stored as the empty string. Empty strings are ignored
in parquet predicate pushdown filtering. So, we need to treat that
separately.

In the same are, we'd strip off the VRF field if some of the rows had
a blank VRF (as is the case with default). Removed this code to ensure
that we always display VRF.