Remove obsolete code that changed '/" in ifnames to '-'
Once upon a time, we used the interface name as a key, which meant
that we had to convert interface names that had "/" in them such as
Ethernet1/1 or xe-0/0/0. This is no longer the case, since release
0.5 or so, and so we no longer need to perform this translation.
Dead code was around in the CLI portion of the code. This patch
removes it.