NXOS's varied return issue strikes again for VLAN
WHen there's just a single entry, NXOS doesn't return a list of
a single element, but it returns the element itself. The code has
to handle this condition. The initial code didn't consider this issue.
The first fix assumed it was always a list, and this one restores
order by checking to see if its a list or a string before operating
on the interfaces data.