interfaces.yml/avsc: add additional fields reqd for EVPN
NXOS and EOS advertise using an anycast gateway mac address in the
case of EVPN routing. This address is not associated with any interface
and is visible via only internal commands (NXOS) or via additional
commands (EOS). In addition, we can also store the route distinguisher
associated with a VRF.

Additional bug fixes associated with gathering interface data across
EOS and NXOS fixed.