Fix ARP/ND entry for Cumulus
In post-4.0 releases, the field that identifies that an ARP/ND entry
is learnt from a remote node has changed. This patch ensures we address
both the pre-4.0 and post-4.0 behaviors of ARP/ND entries for identifying
remotely learnt entries.

We also force the state of the entry to be reachable instead of extern_learn
or anything else in case of remote entries.