Fix #276: Handle username/password that use URL reserved chars
URLs have a set of reserved characters such as '.', ':' and so on.
If a URL of the form <transport>://username:password@hostname:port
uses one of these reserved characters, the inventory for that host
is off. To fix this, we allow users to additionally specify the
username and password separately as keywords. In other words, with
this patch, we support the format:
ssh://10.255.3.250:22 username=vagrant password=foobar

This allows users to have username such as admin.root and still
allow the poller to function. The old format is also supported, but
if the keyword format is specified, that takes precedence over the
value in the old format.