summarize should work with qualified hostnames
Summarize is useful not just for summarizing info about an entire
namespace, but also within specific hostnames. This is useful for
example when the amount of info on a given host is quite a lot, and
the user is looking for summarized info on a specific set of hosts.

This patch also has the test update to REST server testing to account
for this change.

Also accidentally creeping in, is a patch to fix a minor issue with
pager output