engine.py: Handle schema evolution where field name changes
We have a case here where the VLAN resource changed the name of a
field from ifname to vlanName. If you have old data, it'll be under
the old name, and the new data will be under the new name. We need to
ensure filters, drop_duplicates etc. all work correctly with this.
Schema evolution should support this approach.

The code needed migration to handle this. We added a new field called
merge_fields which is a dictionary mapping old field names to new
field names. This is passed by whatever resource needs to do this sort
of rename handling. This specific patch handles the behavior desired
if merge_fields is specified.