test_sqcmds.py: Make the dataframe comparison more robust
Comparing two complex nested data structures for the difference is always
complex. Some of the problems are: different order of fields due to
sorting, fields that are lists have elements in the wrong order, the top
level list being in different sorted order and so on.

We use JSON as the returned output format to compare the expected output
with what was actually returned. Since none of the obvious tools such as
jsondiff worked, we decided to take advantage of pandas itself to see if we
could use pandas dataframe functions to do the diff.

We first started with pandas merge dataframes, and that had a limitation
that caused us to also check the shape of the dataframes to capture an error.
Then, pandas 1.1.0 added support for a true dataframe compare that was much
better. But it didn't handle the case when a column value was a list as the
expected and got could vary in element order. So, we added support to check
that. Next, pandas compare only works when the two dataframes have the same
shape. So, we added an exception handling code to work with pandas merge for
this use case.

But this still left a problem with time fields that could be off a bit because
of the time of collection vs time of collection of test data. This should
largely be fixed by reusing the same test data input to build the parquet as
we do with run-once=gather model, but this doesn't work in some odd cases. So
we added an option to have the test specify such columns to ignore during
compare.

We finally pulled all this complex code into a routine all by itself because
we used it twice, once for comparing failures such as assert, and the other
for normal scenarios.

Thats this patch.