service.py: Don't merge entries of different types
Suzieq supports running multiple commands to extract the data it needs.
If the key used across the commands is the same, it combines records across
the different commands. However, this is sometimes undesirable as in the
case of evpnVni with asymmetric support on Junos. To avoid this, we need to
tell services.py to not combine entries with different types, which is what
this patch does. A separate patch addresses the case of using _entryType as
the field name to use in any service when you don't want the records combined
across the multiple commands.