Fix #241, #179: sq-poller selection of services to run
This patch adds the ability to exclude certain resources from being
monitored by the poller, and also allows the ability to specify
multiple specific resources to be monitored. If both are specified,
the exclude list applies to the list specified as needing to be
monitored.

Rename test files to ospfIf and ospfNbr to reflect consistency