summarize: Accept query string to filter data before summary
With the GUI, its possible to provide a pandas filter to filter the
data shown. To ensure that this is applied uniformly to summarize and
unique, we need to pass the query string to the underlying function in
the engine layer. This patch is to address the issue for summarize.