GUI: New session state logic
The previous session state logic frequently lost track of various
widget positions. To avoid this mess creating user confusion, a
new session state maintenance model has been adopted with this patch.
To avoid losing track, the new session state logic reruns the page
once more. This is inefficient, but with caching, its hopefully less
affected by this.

However, when we maintain old state as in the search results page,
this rerun means everything is redrawn twice which is problematic.
So we ignore rerunning the search page. But this means we have to add
back the old ways of maintaining state when we switch to the search
page to query some more.

All in all, session state maintenance in Streamlit is very complex
and hopefully, native session state maintenance will be in soon.