CLI: Minor change to look for error output
Instead of checking if 'error' is the only column in the output,
we check to see if there's a column called error in the output.
This is useful in mutliple scenarios, the most relevant being the
one with path where we can mark a specific entry with error msg
such as "loop".