poller/interfaces.py: Various fixes for NXOS and JUNOS
This patch addresses two issues:
  - we were not able to efficiently retrieve MTU for switchports for NXOS
  - we missed recording IP addresses on loopback interfaces for JUNOS

For the NXOS issue, the MTU for switchports is only retrievable via "show int"
command which tends to hang as it gathers counters data for VLAN interfaces.
There's no way to get the information for non-Vlan interfaces in a single command
show int quick doesn't do the trick. So, we had to resort to using the ip link
command from the bash prompt to retrieve the MTU for switched ports.

The JunOS bug was a silly accidental line delete during one of the earlier
checkins. Fixed.