poller: Support specifying entry type for textfsm
When we use textfsm to decode a NOS' output, its useful to sometimes
not merge the results of multiple commands. This is true for example in
the case of NXOS' interfaces commandset. In case of JSON'ified output,
its easy to add the entry type info, but not via textfsm. This patch
adds that support.