poller/interfaces: A better attempt at getting NXOS mtu
This command uses textfsm with show int on only ethernet interfaces
to get switchport MTU and macaddr. This fixes the MTU or port channels
too as the MTU of port channels is tied to the underlying ethernet
interface MTU.

We cannot use run bash ip link to retrieve this info for various reasons
including feature enabling bash shell access is not usually enabled and
because we're not sure retrieving the info via the Linux command is tested
and will be assumed to work in future releases.