EOS tests: Using new topology and config for EOS
To maintain a consistent reference base across all NOS, we're using
a new repo with topology and config called automatic-for-the-people.
This patch includes changes to EOS input data and tests to match this
new topology and reference. In future, other NOSes will also be moved
to this new topology and configuration.