Testing NXO: Using new topology and test results
To make testing consistent across all NOS, we've introduced a new
base topology and configuration that all NOS will adhere to. This
patch adds the tests and the input files to generate the parquet
data for that ref topology for NXOS. JunOS and Cumulus will follow
in the next release.

This topology is the 2-spine, 4-leaf, 4 dual-attached servers with
2 exit leaves and a firewall and internet edge router.