path.py: More NXOS behavioral fixes
NXOS prints some unusual interface names in MAC address table. For
example, instead of using the vpc peerlink name, it uses the string
"vPC Peer-Link(R)", and "sup-eth1(R)" for the loopback interface (or
any valid interface to get the packet to the control plane). This
causes path to fail because it doesn't have these interface names in
the device interface database.

This patch fixes that by converting these addresses into something
valid.