interfaces: Ensuring VRF of an interface is set correctly in JUNOS
Due to a bug in the code, subinterfaces (such as xe-0/0/1.1 or irb.10)
were not getting their VRF assigned right. This patch fixes that.

Additionally, the derivation of gateway MAC addr can be from one of two
fields in Junos. We were not extracting the data from one of the fields.
This patch fixes that as well