Fix #304: Fix FastAPI init in code and test
When we added the ability to allow the user to specify the name
and location of the certificate and key files, we accidentally deleted
a critical line which caused the REST server to die. This was setting
an attribute which was the config file. The tests didn't fail because
they explicitly set the attribute file.

To fix this:
- we added a new routine called app_init which takes a config file and sets
  it correctly.
- we make both sq-rest-server and the test routine call this app_init function
- we added a startup function to FastAPI that checks that this config file
  is set. If this is not set, it fails. This way, anybody who tries to import
  and use the app directly will fail.