Support for users specifying the timezone in analysis
Before this patch, the time displayed was in the local timezone of
the CLI/GUI/REST etc. This breaks tests because some developers live
in a different TZ than other developers! To fix this, we decided to
support the specification of a TZ so we could pick a common  TZ for
testing.

The TZ is specifiable in the config file, suzieq-cfg.yml as follows:
analyzer:
  timezone: 'GMT'

Replace GMT with the timezone you prefer. We use GMT for tests. The
format of timezone is 'America/Los_Angeles' or 'Asia/Dilli' or
'Europe/Budapest' or 'NZ'. Its a string accepted by python's pytz
module. Checking the config validity ensures that we can catch any
errors specified by the user. No specification defaults to the local
timezone.