Initial support for IOS-XR
This patch only enables the data gathering part of the process, by
enabling the use of setting up a persistent SSH connection to an IOS-XR
router. The code has been specifically tested against the IOS XRv image
for ASR9K (i.e. VM image).

Pick a static file with a long time for use to maintain a persistent
SSH connection to an IOS router. This ensures that there is no load on
the router CPU as a consequence of this persistent session and also
no network traffic as a result.