Refactoring code: Creating a DB layer
The code had a deficiency in its structure: the engines directory which
did the analysis also included the code to read the data off the DB. What
was engine.py was really the parquet reader. We intend to support other
databases to read the data. So, this patch splits up the code to make
the DB layer a separate thing from the analysis engine.

Deciding which DB is required to be read for which table/resource will be
done in a later release. But we're setup for that, and the analysis engine
needs to know none of these details.