parquet coalescer: Split coalescer bin into front/back end
The coalescer binary contained everything needed to run the coalescer
from the command line. The coalescer also had policy restrictions
such as only one coalescer can be running at any given time. But
this breaks testing because pytest runs several tests in parallel.

To avoid this problem, we split the coalescer binary into a front-end
file that's responsible for all the user-input related work, and a
backend library routine that implements the coalescer run, running
either just once or forever periodically depending on the parameters
passed.

This change also caused the __init__.py routine to change.