sqobjects: Get object by giving table name
The current interface required the user to explicitly invoke the
object necessary by importing it from the appropriate file. For
example, the caller had to know that DeviceObj from suzieq.sqobjects.device
had to be imported to read device tables. This seems onerous.

This patch makes it possible for a user to follow the old logic or
just request the object for servicing a table given its name. So, the
new logic will be get_sqobject('device'). The older model will be most
likely deprecated in time.