Poller: device was writing too often even without change
Whenever we have to rely on timedelta values from a device such as
uptime, to avoid writing a value that changes every poll interval, we
convert relative timestamps into absolute timestamp. For example, we
convert uptime into a bootup absolute time timestamp. However, there
if still the danger of writing an unchanged record too often because
of minor differences in polling interval and therefore the computed
timestamp. In such cases, we either ignore changes to timestamp and
hope we can catch a change via some other method (status field will
change for example with say BGP or OSPF). In case of the device record
we don't have any other field besides the bootupTimestamp to record
that the node has rebooted. In case of device bootup time, we note that
a device has rebooted because the SSH connection resets. There maybe
other reasons an SSH session resets (SSH server restart for example).
But capturing an unchanged value more than once due to a rare issue is
OK because we can filter out these changes in the read/analyze cycle.
In case of bootupTimestamp, we had forgotten to add bootupTimestamp as
a field that had to be ignored.

Fixed this by adding bootupTimestamp as a field to be ignored when
checking if a record has changed.