Modifications to unique
unique() was in reality displaying value counts which represented counts
that weren't always easy to derive understanding of except that it was
the count of rows with that value. In some cases, the count made a lot
more sense than in other case.

To avoid this confusion as well as make unique be how unique is interpreted
in other analysis tools, we changed unique to be just a list of unique
values. An additional attribute, count=True produces value counts.