Support new option to store keys in second level dict
In FRR's BGP output, you have entries that look like this:
  "multiprotocolExtensions":{
        "ipv4Unicast":{
          "advertisedAndReceived":true
        },
        "ipv4Multicast":{
          "received":true
        },

And we want to save all AFI/SAFI that are not advertisedAndReceived.
In the current model there's no way to save the key of ipv4Multicast
or ipv4Unicast, without saving the entire multiprotocolExtensions and
processing each key in python. Instead, we added the option called
_sqstore which should have a low chance of conflict and its used as
"multiprotocolExtensions/*:_sqstore/received: afisRcvOnly". This is
used to indicate that we need to save the value of the key in
afisRcvOnly, not the value of received.