Add support for augmented columns in tables
Augmented columns are virtual columns in that their contents are
computed from other fields in the table. For example, prefixlen is
not a column in routes. By defining prefixlen as an augmented column,
we can now compute prefixlen from the column prefix, and prefixlen can
be used in show, filtering, unique etc. Augmented columns are not
written out to the backend store, but computed at the front end due
to analysis.

An augmented column is defined in the schema via an additional "depends"
clause in its definition.

In this initial version, there's no support for augmented columns that
are derived from columns in a different table. All augmented columns
MUST depends only on other columns in the same table.