OSPF: make query_str processing uniform across tables
OSPF's handling of user-specified pandas query string was not handled
the same way as the other tables which caused all sorts of errors.
This patch makes it uniform with the others.

However the behavior of all tables in handling the query string needs
to be modified with the addition of augmented columns. This'll come
next.