Fix #308: Provide password via an environment variable
A common way to pass passwords in K8s and other environments is to
supply it via an environment variable. This prevents leaving a password
in either clear text in the inventory file or having to type it manually
every time the poller starts.

This patch supports this by adding a new option to the poller,
envpass, which provides the name of the environment variable to read the
password in from. So, sq-poller -D inventory.yml -k --envpass=SQPASS,
means read the password from the environment variable called SQPASS. By
making the environment variable specific to the user, any attempt to read
a well-known env var is also prevented.

If the user specifies the password in the inventory file for a specific set of
devices, that overrides the environment variable password for those devices.