Reorganzing dirs to make it suitable as a PyPi package
To make Suzieq installable as a regular Python package, some reorganization
of the directory structure was required. This patch addresses this.

In addition to reoorganizing, we also made finding the service and schema
dirs automatic by the code rather than via specification in the config file
as these dirs are now hidden behind wherever the installation happens.

Also made the rest server only bind to 127.0.0.1 and made the bind address
and port configurable.