Device: Add devices whose poller has failed to dataframe
Before this patch, devices that could not be polled just did not show
up in the Device table outputs. The users had to go look at the
sqpoller table output to determine this. With this patch, we fix that.
Devices that have never had a single poll on them show up as 'neverpoll'
whereas devices that had been polled but have since stopped responding
show up as 'dead'