Fix #363: Accept mac address in various formats in query
Users habituated to using macaddr in certain ways such as the
dotted notation used on Cisco devices (1111.2222.3333) or used to
seocifying capital letters for a-f in macaddr will be uncomfortable
with the only format Suzieq accepts today, the standard mac addr
format with 6 digit-pairs separated by ':' and a-f being in lowercase.

This patch addresses the issue by creating a new dictionary that is
used to massage the input into the internal format accepted by Suzieq.
Take a look at suzieq/sqobject/macs.py or arpnd.py for examples.

The issue is tricky in case of address table where the mac address can
be buried inside the address string filter. So, we handle this one
differently as coded up in engines/pandas/address.py