Fix getting correct files to read for a time period
In the presence of coalesced files, we need to pick the right files
to read for a given time period or when looking for the latest info.
There were a few bugs in getting this file list right, and the existing
code was also inefficient. This patch fixes that.

In addition, an unnecessary addon to this is the fixups to support
coalescing periods in the order of minutes.