BGP: Vastly improved determining of peerHostname
In the presence of large amounts of data, finding the peerhostname
would sometimes never complete, or work very slowly due to the huge
amount of memory consumed by the existing algorithm. For example,
in one case involving 90K BGP records, the code consumed 50GB of
memory and took a long time to complete.

This patch makes that logic far more efficient. The same 90K of
BGP records now completes in under a second and with < 2GB of memory
consumed.