Coalescer: Handle removing dup index correctly
In coalescing, we retain the records for the current state of all keys
so that we can access the info about any record in constant time. In
doing this, we ended up having the old numerical index field called "index"
stick around and this caused a problem in certain scenarios by getting
rid of certain records.

This patch addresses the issue by dropping the index field and using the
correct index fields to dedup.