Poller, files: Make command independent of format
When we're dealing with devices that accept REST as well as SSH,
the command's maybe different based on the requested format. For
example, we add "| json" if we're communicating with an EOS router.
This messes up the ability to serve the right output file. We
therefore standardize the command without the filters (anything after
a | if present) when we create the dictionary of command to output.