Update critical libraries
This updates pandas version to 1.2.5, pyarrow to 5.0.0, and streamlit
to 0.86. Consequently, the update to pyarrow causes a problem with a
specific code line in the codebase.

Using "if or_filters" when or_filters is a Pyarrow expression is not
allowed. We need to explicit replace the check with "if or_filters is not None"
This logic is in parquetdb.py and so this commit also contains this fix.