get_sqobject: Return the right object for all tables
get_sqobject's logic for determining if a member of a module was
the right object for a class was broken. As a consequence you couldn't
get the right object for tables like path.

This fix ensures that the object we return belongs to the parent class
SqObject. It also raises a ModuleNotFoundError exception if such a
class cannot be found for the given service name.