Fix #307: Add ability to log to stdout
Running suzieq within K8s requires logging to stdout to make it fit with
K8s infra and setup. Even without K8s, logging to stdout maybe interesting
for many users, especially when starting off.

This patch adds the ability to log to stdout. You do this by adding the
keyword log-stdout and assign it to True. In this model, logging goes to
both the logfile and to stdout. If you don't want to log to stdout, define
the logfile parameter in the suzieq-cfg file and leave it blank. If the logfile
parameter is not specified, its assumed to go to the default log file associated
with the facility.