Handle changes due to streamlit session state support
This also fixes the issue of streamlit now using Arrow to serialize the
data for what is claimed to be a much superior performance. Unfortunately,
the way we construct our summary dataframe, a single column has a bunch of
different data types that breaks the Arrow serialization. The simplest fix
is to convert the dtype of the summary_df to str before passing to
st.datatable. This is the least disruptive method to allow us to use the
new serialization format, rather than rely on the legacy one. We'll fix the
summary dataframe correctly in a future release.