GUI: Handle macaddr input forms correctly
The user can specify a MAC address in any format: the traditional
6-byte model with each octet separated by ':', or the Cisco style
where 3 2-byte values separated by '.'. Further they can use either
upper case or lower case in the mac address as is the norm for the
platform.

We were not handling all these cases and the fact that you can
the Cisco format for tables such as ARPND where the user can specify
either the IP address or the MAC address.

This patch addresses all that