CLI: Add help verb to provide better help
There was no real help available via the CLI for someone who was starting
off with Suzieq. help just provided a one line help summary, nothing more.
To that end, this patch adds a help verb that provides at the top level,
a help with additional text as to the verbs supported for each table/resource.
A second level of help is to specify the specific subcommand(or verb) and get
additional help for that verb. Right now, everything is based off of
docstrings, and so it was required to cleanup some of the docstrings of
functions. Moving forward, we can add additional help such as the checks that
are performed for each assert, choices for each argument, if any and so on.

Examples:

network help
network help command=find