CLI/command: Add an additiional vars help param
For whatever reason, classargspec doesn't provide info about the vars
that have been declared only for a specific object. For example, path
has src/dest/vrf defined as being applicable to all verbs of that object.
But, classargspec has not info about these vars, and as a consequence,
the help for any of the verbs of path do not display these three critical
vars.

To help with this, we define a new class ArgHelpClass that contains the
fields we extract from a classargspec to print the help. We then define
a new class param called _additional_help_vars as a dictionary to carry
the help for these class-specific vars. Path is the only user for now.