Add pylint linter to pre-commit
This has most of the code files excluded from pylint since many files
currently do not pass but sets up the possibility for the project to
start updating files to conform to pylint. The excludes list can be
updated as this happens to start running pylint checks on said files.

Also adds pydoc strings to a few files missing them and resolve
pylint's W0107 error.

Ref: https://vald-phoenix.github.io/pylint-errors/plerr/errors/basic/W0107.html
Signed-off-by: Thanh Ha <zxiiro@gmail.com>