Fix var 'cfgfile' referenced before assignment
The function sq_get_config_file() appears to intend to return cfgfile
regardless of if it is found or not as the load_sq_config() function
appears to have a condition to exit if a cfgfile is not present.

This resolves the "UnboundLocalError: local variable 'cfgfile'
referenced before assignment" thrown when trying to start the
suzieq-gui without a config file present to allow the sys.exit(1)
in load_sq_config() to exit.

Signed-off-by: Thanh Ha <zxiiro@gmail.com>